Settings page and routes refactor
