+++ src/locales/en/common.json
@@ -95,5 +95,6 @@
         "ph": "Tagalog",
         "cb": "Cebuano"
     },
-    "noResults": "No results found"
+    "noResults": "No results found",
+    "settings": "Settings"
 }
