+++ src/locales/en/pages.json
@@ -76,7 +76,9 @@
     "dappConnectionsSub": "Forget all of the dapps connected.",
     "done": "Done!",
     "forgetAllConnections": "Forget all connections",
-    "version": "Version"
+    "version": "Version",
+    "accounts": "Accounts",
+    "accountsSub": "Manage account default network settings."
   },
   "experiments": {
     "experiments": "Experiments",
