+++ src/views/AccountSettings.vue
@@ -1,0 +1,8 @@
+<template>
+  <h2>Account settings</h2>
+</template>
+<script>
+export default {
+  name: 'AccountSettings'
+}
+</script>
